+++ src/lib.rs
@@ -20,7 +20,7 @@ Instead of reacting to raw input data like "Released" or "Pressed", the crate pr
 [`Modifiers`](input_context::input_modifier) let you change the input before passing it to the action. We provide common modifiers,
 like [`DeadZone`], [`Negate`], etc., but you can add your own by implementing [`InputModifier`].
 
-[`Conditions`](input_context::input_condition) define how an action activates. We also provide built-in conditions, such as [`Pressed`],
+[`Conditions`](input_context::input_condition) define how an action activates. We also provide built-in conditions, such as [`Press`],
 [`Release`], [`Hold`], etc. You can also add your own by implementing [`InputCondition`].
 
 # Quick start
@@ -72,7 +72,7 @@ pub mod prelude {
             input_action::{Accumulation, ActionEvent, ActionEventKind, ActionState, InputAction},
             input_condition::{
                 blocked_by::*, chord::*, condition_timer::*, down::*, hold::*, hold_and_release::*,
-                pressed::*, pulse::*, release::*, tap::*, ConditionKind, InputCondition,
+                press::*, pulse::*, release::*, tap::*, ConditionKind, InputCondition,
             },
             input_modifier::{
                 dead_zone::*, delta_lerp::*, delta_scale::*, exponential_curve::*, negate::*,