+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Rename `ActionBind::with_axis2d` into `ActionBind::with_xy_axis`.
 - Rename `ScaleByDelta` into `DeltaScale`.
 - Rename `Released` into `Release`.
+- Rename `Pressed` into `Press`.
 - Replace `SmoothDelta` with `LerpDelta` that does only linear interpolation. Using easing functions for inputs doesn't make much sense.
 - Modifiers are now allowed to change passed value dimensions.
 - All built-in modifiers now handle values of any dimention.